[Refactoring] Moving REST resources and components to internal packages.
