+++ xwiki-enterprise-test/xwiki-enterprise-test-rest/src/test/it/org/xwiki/test/rest/CommentsResourceTest.java
@@ -28,13 +28,13 @@ import org.apache.commons.httpclient.methods.PostMethod;
 import org.junit.Assert;
 import org.junit.Test;
 import org.xwiki.rest.Relations;
+import org.xwiki.rest.internal.resources.comments.CommentsResource;
+import org.xwiki.rest.internal.resources.pages.PageHistoryResource;
 import org.xwiki.rest.model.jaxb.Comment;
 import org.xwiki.rest.model.jaxb.Comments;
 import org.xwiki.rest.model.jaxb.History;
 import org.xwiki.rest.model.jaxb.HistorySummary;
 import org.xwiki.rest.model.jaxb.Page;
-import org.xwiki.rest.resources.comments.CommentsResource;
-import org.xwiki.rest.resources.pages.PageHistoryResource;
 import org.xwiki.test.rest.framework.AbstractHttpTest;
 
 public class CommentsResourceTest extends AbstractHttpTest
