+++ xwiki-enterprise-test/xwiki-enterprise-test-rest/src/test/it/org/xwiki/test/rest/PageResourceTest.java
@@ -33,6 +33,12 @@ import org.apache.commons.httpclient.methods.PutMethod;
 import org.junit.Assert;
 import org.junit.Test;
 import org.xwiki.rest.Relations;
+import org.xwiki.rest.internal.resources.SyntaxesResource;
+import org.xwiki.rest.internal.resources.pages.PageChildrenResource;
+import org.xwiki.rest.internal.resources.pages.PageHistoryResource;
+import org.xwiki.rest.internal.resources.pages.PageResource;
+import org.xwiki.rest.internal.resources.pages.PageTranslationResource;
+import org.xwiki.rest.internal.resources.wikis.WikisResource;
 import org.xwiki.rest.model.jaxb.History;
 import org.xwiki.rest.model.jaxb.HistorySummary;
 import org.xwiki.rest.model.jaxb.Link;
@@ -45,12 +51,6 @@ import org.xwiki.rest.model.jaxb.Syntaxes;
 import org.xwiki.rest.model.jaxb.Translation;
 import org.xwiki.rest.model.jaxb.Wiki;
 import org.xwiki.rest.model.jaxb.Wikis;
-import org.xwiki.rest.resources.SyntaxesResource;
-import org.xwiki.rest.resources.pages.PageChildrenResource;
-import org.xwiki.rest.resources.pages.PageHistoryResource;
-import org.xwiki.rest.resources.pages.PageResource;
-import org.xwiki.rest.resources.pages.PageTranslationResource;
-import org.xwiki.rest.resources.wikis.WikisResource;
 import org.xwiki.test.rest.framework.AbstractHttpTest;
 import org.xwiki.test.rest.framework.TestConstants;
