+++ xwiki-enterprise-test/xwiki-enterprise-test-rest/src/test/it/org/xwiki/test/rest/RootResourceTest.java
@@ -24,9 +24,9 @@ import org.apache.commons.httpclient.methods.GetMethod;
 import org.junit.Assert;
 import org.junit.Test;
 import org.xwiki.rest.Relations;
+import org.xwiki.rest.internal.resources.RootResource;
 import org.xwiki.rest.model.jaxb.Link;
 import org.xwiki.rest.model.jaxb.Xwiki;
-import org.xwiki.rest.resources.RootResource;
 import org.xwiki.test.rest.framework.AbstractHttpTest;
 
 public class RootResourceTest extends AbstractHttpTest
