+++ xwiki-enterprise-test/xwiki-enterprise-test-rest/src/test/it/org/xwiki/test/rest/SpacesResourceTest.java
@@ -24,6 +24,10 @@ import org.apache.commons.httpclient.methods.GetMethod;
 import org.junit.Assert;
 import org.junit.Test;
 import org.xwiki.rest.Relations;
+import org.xwiki.rest.internal.resources.spaces.SpaceAttachmentsResource;
+import org.xwiki.rest.internal.resources.spaces.SpaceSearchResource;
+import org.xwiki.rest.internal.resources.wikis.WikiSearchResource;
+import org.xwiki.rest.internal.resources.wikis.WikisResource;
 import org.xwiki.rest.model.jaxb.Attachment;
 import org.xwiki.rest.model.jaxb.Attachments;
 import org.xwiki.rest.model.jaxb.Link;
@@ -33,10 +37,6 @@ import org.xwiki.rest.model.jaxb.Space;
 import org.xwiki.rest.model.jaxb.Spaces;
 import org.xwiki.rest.model.jaxb.Wiki;
 import org.xwiki.rest.model.jaxb.Wikis;
-import org.xwiki.rest.resources.spaces.SpaceAttachmentsResource;
-import org.xwiki.rest.resources.spaces.SpaceSearchResource;
-import org.xwiki.rest.resources.wikis.WikiSearchResource;
-import org.xwiki.rest.resources.wikis.WikisResource;
 import org.xwiki.test.rest.framework.AbstractHttpTest;
 
 public class SpacesResourceTest extends AbstractHttpTest
