+++ xwiki-enterprise-test/xwiki-enterprise-test-rest/src/test/it/org/xwiki/test/rest/TagsResourceTest.java
@@ -31,16 +31,16 @@ import org.apache.commons.httpclient.methods.PutMethod;
 import org.junit.Assert;
 import org.junit.Test;
 import org.xwiki.rest.Relations;
+import org.xwiki.rest.internal.resources.pages.PageResource;
+import org.xwiki.rest.internal.resources.pages.PageTagsResource;
+import org.xwiki.rest.internal.resources.tags.PagesForTagsResource;
+import org.xwiki.rest.internal.resources.tags.TagsResource;
 import org.xwiki.rest.model.jaxb.Link;
 import org.xwiki.rest.model.jaxb.Page;
 import org.xwiki.rest.model.jaxb.PageSummary;
 import org.xwiki.rest.model.jaxb.Pages;
 import org.xwiki.rest.model.jaxb.Tag;
 import org.xwiki.rest.model.jaxb.Tags;
-import org.xwiki.rest.resources.pages.PageResource;
-import org.xwiki.rest.resources.pages.PageTagsResource;
-import org.xwiki.rest.resources.tags.PagesForTagsResource;
-import org.xwiki.rest.resources.tags.TagsResource;
 import org.xwiki.test.rest.framework.AbstractHttpTest;
 import org.xwiki.test.rest.framework.TestConstants;
